Split MarkdownCommitMessageBuilder file
