+++ formatter/src/commonMain/kotlin/builder/markdown/MarkdownCommitMessageBuilder.kt
@@ -10,8 +10,7 @@ package builder.markdown
 
 import builder.CommitMessageBuilder
 import builder.FormattedCharSequence
-import builder.markdown.TokenType.Comment
-import builder.markdown.TokenType.MarkdownToken
+import builder.markdown.TokenType.GitComment
 import builder.markdown.TokenType.MarkdownToken.*
 
 internal data class MarkdownCommitMessageBuilder(
@@ -49,84 +48,12 @@ internal data class MarkdownCommitMessageBuilder(
                 EndPunctuation -> sequence.appendRaw(token.text)
                 StandalonePunctuation -> sequence.append(token.text.trim())
                 ParagraphBreak -> sequence.breakParagraph()
-                is Comment,
+                is GitComment,
                 WordSpacing,
                 SingleLineBreak -> {
                 }
             }
             currentPosition += token.text.length
         }
-    }
-}
-
-private data class Tokenizer(
-    private val commentChar: Char,
-) {
-
-    private val tokenTypes = arrayOf(
-        Comment(commentChar),
-        *MarkdownToken.values(),
-    )
-
-    fun tokenizePosition(
-        currentPosition: Int,
-        body: String,
-    ): Token {
-        for (type in tokenTypes) {
-            val match = type.matchAt(body, currentPosition) ?: continue
-            return Token(type, match)
-        }
-        val currentPositionSnippet = body.substring(currentPosition, currentPosition + 10)
-        error("No match for position $currentPosition: $currentPositionSnippet...")
-    }
-
-    @OptIn(ExperimentalStdlibApi::class)
-    private fun TokenType.matchAt(
-        string: String,
-        index: Int,
-    ): String? {
-        val match = Regex(pattern).matchAt(string, index)
-        return match?.value
-    }
-
-    data class Token(
-        val type: TokenType,
-        val text: String,
-    )
-}
-
-private sealed interface TokenType {
-
-    val pattern: String
-
-    data class Comment(
-        val commentChar: Char,
-    ) : TokenType {
-        //language=RegExp
-        override val pattern = "$commentChar.*"
-    }
-
-    @Suppress("RegExpRedundantEscape") // Some are needed for JS
-    enum class MarkdownToken(
-        //language=RegExp
-        override val pattern: String,
-    ) : TokenType {
-        ParagraphBreak("""\n\s*\n"""),
-        Heading("""#+[^\n]+"""),
-        BoldText("""\*\*[^*]+\*\*"""),
-        ItalicizedText("""_[^_]+_"""),
-        Strikethrough("""~[^~]+~"""),
-        Quote(""">[^\n]+(?:\n>[^\n]+)*"""),
-        Table("""\|.*\|(?:\n\|.*\|){2,}"""),
-        BulletListItem("""[ \t]*- [^\n]+(?:\n[^\n]+|[ \t]*- [^\n]+)*"""),
-        NumberedListItem("""[\t ]*\d{1,9}\. [^\n]+(?:\n[^\n]+|\n[\t ]*\d{1,9}\. [^\n]+)*"""),
-        CodeSnippet("""```[^`]+```"""),
-        HtmlElement("""<[^\/>]+\/>|<(\w+)[^\/>]*>[\s\S]*?<\/\1>"""),
-        Link("""!?\[[^\]]+\][\[(][^\])]+[\])]"""),
-        PlainText("""(?:\S*\w+\S*)+"""),
-        StandalonePunctuation("""[\t ]+[^\w\s]+(?=\s)"""),
-        WordSpacing(" +"),
-        SingleLineBreak("""\n"""),
-        EndPunctuation("""\s*[^\w\s](?=$|\n\n)"""),
     }
 }
